requirements.txt
flask==3.1.0
mysql-connector-python==8.0.27